feat(engine): widen loop and beat-jump beats to f32
Allow fractional beat sizes on the wire and in controller map args
(e.g. auto_loop(beats:0.25)), closing #137. Rational 1/4 grammar stays deferred.

+++ crates/engine-api/src/payload.rs
@@ -230,16 +230,16 @@ pub enum CmdBody {
         enabled: bool,
     },
     SetAutoLoop {
-        beats: u32,
+        beats: f32,
     },
     BeatJump {
-        beats: i32,
+        beats: f32,
     },
     SetPadMode {
         mode: PadMode,
     },
     BeginLoopRoll {
-        beats: u32,
+        beats: f32,
     },
     TriggerHotCue {
         position_ms: i32,
